Fix mc/1.21.1: NeoForge 21.1.1, SoundEvent registry, CustomModelData int cast
- Correct neo_version to 21.1.1 (21.1.0-beta does not exist on maven)
- BuiltInRegistries.SOUND_EVENT.get() returns nullable SoundEvent in 1.21.1, not Optional<Holder>
- CustomModelData single-arg constructor takes int, not float

+++ src/main/java/moe/feo/shootexp/item/ExpItem.java
@@ -65,9 +65,9 @@ public final class ExpItem {
             stack.set(DataComponents.CUSTOM_MODEL_DATA,
                     (CustomModelData) ctor.newInstance(List.of(value), List.of(), List.of(), List.of()));
         } catch (NoSuchMethodException e) {
-            stack.set(DataComponents.CUSTOM_MODEL_DATA, new CustomModelData(value));
+            stack.set(DataComponents.CUSTOM_MODEL_DATA, new CustomModelData((int) value));
         } catch (Exception e) {
-            stack.set(DataComponents.CUSTOM_MODEL_DATA, new CustomModelData(value));
+            stack.set(DataComponents.CUSTOM_MODEL_DATA, new CustomModelData((int) value));
         }
     }
